Add test for CorePluginsRepository
Using dependency on release.war in Bazel junit_test rule, using data
attribute, we can depend on release.war content from the test: [1].

Then we can retrieve the TEST_SRCDIR passed through environment variable
to the test and retrieve the content of the release.war from within the
tests. The only disadvantage of this approach, is that the test in this
form wouldn't work from within IDE. But in this case release.war could
be built manually and passed in the test by other means.

[1] https://docs.bazel.build/versions/master/build-ref.html#data

Change-Id: I17128017e2ec0287baf34e84b35eed1683f3f422