PluginsRepositoryTest: Explicitly assert for expected env variables
If either the TEST_SRCDIR or TEST_WORKSPACE environment variables is
not defined, the subsequent call to Paths.get will fail with NPE on
dereferencing the null value returned by System.getenv.

Refactor it so that the values are explicitly asserted to be non-null
prior to dereferencing, so that it fails with a meaningful message.

Also refactor the subsqeuent check for existence of the release.war to
use an assertion with meaningful failure message, rather than throwing
an IllegalStateException.

Change-Id: Ibf9463cc237bd0f613a09a093b5257af92e7efda