Add Gerrit logo to the intro page
After 12 years of the Gerrit Code Review project
add the logo to the splash page [1] displayed in the
out of the box experience of Gerrit.

By showing the diffy logo on the very first page
the new users see when they first run Gerrit, the
product gives the message of a renewed attention
to the design and user-experience of modern
Gerrit v3.

[1] https://imgur.com/5gpumX8

Change-Id: I6e7085759153c584350ce41e56328298f4f1954c